Document for damage.py
Done adding inline documentation for damage.py. Three functions need the
new paper to be understand. The have new function for damge which is
relevnet to forcing. And focing is also relevent to the path (definition
is needed in the paper) And the interpolate coffeificent matrix for the
damge function is need to be defined mathmatically.